requirements.txt
asgiref==3.8.1
Django==5.0.4
sqlparse==0.4.4
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.25.0
